% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honeycomb.R
\name{honeycomb}
\alias{honeycomb}
\title{Honeycomb}
\usage{
honeycomb(
  hex_size = 50L,
  hex_stroke_width = 6L,
  flash_color = "#008080",
  basis_color = "#DCDCDC",
  flash_duration_secs = 6L,
  flash_interval_secs = 1L,
  ...,
  sizing = htmlwidgets::sizingPolicy(padding = 0)
)
}
\arguments{
\item{hex_size}{scale of the hexes. Default is 50.}

\item{hex_stroke_width}{width of the stroke dividing the hexes}

\item{flash_color}{hex color of the flashing hexes}

\item{basis_color}{basic color of the unflashing hexes}

\item{flash_duration_secs}{seconds duration of the flashing}

\item{flash_interval_secs}{seconds interval of each trigger}

\item{...}{passed to \code{\link[r2d3:r2d3]{r2d3::r2d3()}}}

\item{sizing}{sizing policy according to \code{\link[htmlwidgets:sizingPolicy]{htmlwidgets::sizingPolicy()}}}
}
\description{
A hexmap with randomly flashing hexes
}
